refactor: [YSS-72] 패키지 오타 변경
