+++ yakssok/src/main/java/server/yakssok/domain/feedback/domain/repository/FeedbackRepository.java
@@ -1,8 +1,8 @@
-package server.yakssok.domain.feeback.domain.repository;
+package server.yakssok.domain.feedback.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import server.yakssok.domain.feeback.domain.entity.Feedback;
+import server.yakssok.domain.feedback.domain.entity.Feedback;
 
 public interface FeedbackRepository extends JpaRepository<Feedback, Long>, FeedbackQueryRepository {
 }
